+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * minor pylint improvements
 * better coverage
 * SyncableDroppable.cleanup() bugs out, wrote a quick patch for it to do nothing and filed as #61.
+* unit tests for overloading